requirements.txt
# =============================================================================
# SetuPranali - Python Dependencies
# =============================================================================
# Install: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# -----------------------------------------------------------------------------
# Database Drivers
# -----------------------------------------------------------------------------
duckdb>=0.9.0
psycopg2-binary>=2.9.0

# -----------------------------------------------------------------------------
# Cloud Data Warehouse Drivers (Optional - install as needed)
# -----------------------------------------------------------------------------
# Snowflake (TB-scale analytics)
snowflake-connector-python>=3.6.0

# BigQuery (TB-scale analytics on GCP)
google-cloud-bigquery>=3.0.0
google-cloud-bigquery-storage>=2.0.0

# Databricks (TB-scale analytics on lakehouse)
databricks-sql-connector>=3.0.0

# ClickHouse (real-time analytics)
clickhouse-connect>=0.7.0

# Redshift (AWS data warehouse)
redshift-connector>=2.0.0

# MySQL / MariaDB
mysql-connector-python>=8.0.0

# -----------------------------------------------------------------------------
# Security & Encryption
# -----------------------------------------------------------------------------
cryptography>=41.0.0

# -----------------------------------------------------------------------------
# Caching & Rate Limiting
# -----------------------------------------------------------------------------
redis>=5.0.0
slowapi>=0.1.9

# -----------------------------------------------------------------------------
# Configuration
# -----------------------------------------------------------------------------
pyyaml>=6.0.0
python-dotenv>=1.0.0

# -----------------------------------------------------------------------------
# HTTP Client (for health checks in WDC)
# -----------------------------------------------------------------------------
httpx>=0.25.0

